requirements.txt
# Azure
azure-storage-blob==12.16.0
azure-identity==1.13.0
azure-core==1.26.4

# Data processing
pandas==2.0.3
openpyxl==3.1.2

# LLM and NLP
langchain==0.0.235
requests==2.31.0

# UI and visualization
streamlit==1.24.0
plotly==5.15.0

# Testing
pytest==7.4.0

# Other utilities
python-dotenv==1.0.0